changed approach to single handle fetching sensor data

+++ src/main/java/skni/kamilG/skin_sensors_api/Sensor/Model/SensorData.java
@@ -25,7 +25,7 @@ public class SensorData {
 
   private LocalDateTime timestamp;
 
-  private Short temperature;
+  private Float temperature;
 
   private Integer humidity;
